исправить ошибки
