+++ .eslintrc
@@ -1,10 +1,6 @@
 {
   "extends": "airbnb-base",
   "rules": {
-    "no-underscore-dangle": ["error", { "allow": ["_id"]}],
-    "no-console": "off",
-    "max-classes-per-file": "off",
-    "no-unused-vars": "off"
-    // "no-else-return": "off"
+    "no-underscore-dangle": ["error", { "allow": ["_id"]}]
     }
 }
