+++ app.js
@@ -12,9 +12,6 @@ const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 const app = express();
 
-console.log(process.env.NODE_ENV);
-console.log(process.env.DB);
-
 const { PORT = 3000 } = process.env;
 
 app.use(bodyParser.json());
@@ -23,10 +20,10 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(requestLogger); // подключаем логгер запросов
 
-app.use((req, res, next) => {
-  console.log(req.method, req.path, req.params);
-  next();
-});
+// app.use((req, res, next) => {
+//   console.log(req.method, req.path, req.params);
+//   next();
+// });
 
 app.post('/api/signup', validateSignup, createUser);
 app.post('/api/signin', validateLogin, login);
@@ -37,7 +34,7 @@ app.use('/api', require('./routes/user'));
 
 app.use('/api', require('./routes/movie'));
 
-app.use((req, res) => {
+app.use(() => {
   throw new NotFoundError('Путь не найден');
 });
 
@@ -48,14 +45,12 @@ app.use(errors()); // это обработчик ошибок celebrate
 app.use(errorHandler);
 
 async function main() {
-  await mongoose.connect('mongodb://0.0.0.0:27017/bitfilmsdb', 
-  {
+  await mongoose.connect('mongodb://0.0.0.0:27017/moviesdb', {
     useNewUrlParser: true,
     useUnifiedTopology: true,
   });
 
   await app.listen(PORT);
-  console.log(`server listen on ${PORT}`);
 }
 
 main();
