+++ errors/doubleUseEmailError.js
@@ -1,0 +1,8 @@
+class DoubleUseEmailError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 409;
+  }
+}
+
+module.exports = DoubleUseEmailError;
