+++ errors/errors.js
@@ -1,43 +1,13 @@
-class BadRequestError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 400;
-  }
-}
-
-// проблемы с токеном и с авторизацией
-class UnauthorizedError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 401;
-  }
-}
-
-class NoRightsError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 403;
-  }
-}
-
-class NotFoundError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 404;
-  }
-}
-
-class DoubleUseEmailError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 409;
-  }
-}
+const BadRequestError = require('./badRequestError');
+const DoubleUseEmailError = require('./doubleUseEmailError');
+const NoRightsError = require('./noRightsError');
+const NotFoundError = require('./notFoundError');
+const UnauthorizedError = require('./unauthorizedError');
 
 module.exports = {
-  NotFoundError,
   BadRequestError,
   DoubleUseEmailError,
-  UnauthorizedError,
   NoRightsError,
+  NotFoundError,
+  UnauthorizedError,
 };
