+++ errors/noRightsError.js
@@ -1,0 +1,8 @@
+class NoRightsError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 403;
+  }
+}
+
+module.exports = NoRightsError;
