+++ errors/unauthorizedError.js
@@ -1,0 +1,9 @@
+// проблемы с токеном и с авторизацией
+class UnauthorizedError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 401;
+  }
+}
+
+module.exports = UnauthorizedError;
