+++ middlewares/auth.js
@@ -3,7 +3,6 @@ const jwt = require('jsonwebtoken');
 const { UnauthorizedError } = require('../errors/errors');
 
 const { NODE_ENV, JWT_SECRET } = process.env;
-console.log(NODE_ENV, JWT_SECRET, 'файл auth');
 
 module.exports = (req, res, next) => {
   const { authorization } = req.headers;
