+++ middlewares/errorHandler.js
@@ -1,11 +1,10 @@
 const errorHandler = (err, req, res, next) => {
   const status = err.statusCode || 500;
-  console.log(err.stack || err);
 
   res.status(status).send({
-    err,
     message: err.message,
   });
+  next();
 };
 
 module.exports = errorHandler;
