+++ middlewares/validation.js
@@ -2,12 +2,6 @@ const { celebrate, Joi } = require('celebrate');
 const { ObjectId } = require('mongoose').Types;
 const validator = require('validator');
 
-// const validateUserId = celebrate({
-//   params: Joi.object().keys({
-//     userId: Joi.string().hex().length(24).required(),
-//   }),
-// });
-
 const validateMovieId = celebrate({
   params: Joi.object().keys({
     id: Joi.string().required().custom((value, helpers) => {
@@ -46,7 +40,7 @@ const validateMovieData = celebrate({
       }
       return helpers.message('Невалидная ссылка');
     }),
-    id: Joi.number().required(),
+    id: Joi.number(),
   }),
 });
 
@@ -85,7 +79,7 @@ const validateSignup = celebrate({
   body: Joi.object().keys({
     email: Joi.string().required().email(),
     password: Joi.string().required(),
-    name: Joi.string().min(2).max(30),
+    name: Joi.string().required().min(2).max(30),
   }),
 });
