+++ package.json
@@ -25,11 +25,11 @@
     "body-parser": "^1.20.0",
     "celebrate": "^15.0.1",
     "dotenv": "^16.0.0",
+    "eslint": "^8.14.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
     "express": "^4.17.3",
     "express-winston": "^4.2.0",
-    "joi": "^17.6.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.3.1",
     "mongoose-unique-validator": "^3.0.0",
